Comment out "checkOwners" since it doesn't work expectedly now
It's possible to override "image" with any docker images with violation because now the validation admission webhook for Deployment doesn't check "UPDATE" operation, then Kritis allows any docker images in ReplicaSet and Pod creation because of the "checkOwners".

If we add "UPDATE" operation to the webhook for Deployment, the webhook can get AdmissionReview for Deployment update but it cause another issue which we can't know the admission comes from "UPDATE" or "DELETE" operation. (Both sends "UPDATE" admission)
It's probably because the admission's Kind is "extensions/v1beta1" and it doesn't support "DELETE" operation.

So for now it's better to disable the "checkOwners" so that we can reject docker images when Deployment updated.

---
The current issue is, it is possible to change docker image of a pod using kubectl edit pod command to any docker image. Validation webhook will skip new image validation check because it will first try to check its owner (Deployment) image. When a pod image is edited to a different image, it does not change Deployment image. So, Deployment image is still same (old one) which is already passed by validation webhook. And validation webhook will let it go.

For now it make sense to just diable owner check functionality. It does save some time when there are many pods of same deployment. In that case it will only check once instead of checking for each pod. It is good for optimization but there is a security loophole.