Interpret NoteReference as full Note name, not just the project id.
This PR changes the way attestation authority's NoteReference field is
interpreted.  It is now required to be of the form "projects/*/notes/*"
and represents the full name of the Note under which the aa's
attestations are stored.  This name can now vary independently from the
name of the attestation authority.

Also in this PR:
Bug fix: use ListNoteOccurrences in grafeas.go instead of
ListOccurrences.  (If some aa's attestations are scattered across
different projects, then ListOccurrences won't return all of them.)